Hotfix

+++ Bermuda/Bermuda/Game.cpp
@@ -40,14 +40,15 @@ void Game::gameLoop(GameStateManager* gsm) {
 	QueryPerformanceCounter(&previousTime);
 
 	//Instantiate FPS variables and set to 0
-	long lastFpsTime = 0;
+	float lastFpsTime = 0;
 	int fps = 0;
 	int gameState = 0;
 	while (gsm->running()) {		
 		//Take current Time
 		QueryPerformanceCounter(&currentTime);
 		//Calculate difference (previousTime - currentTime)
-		long updateLength = static_cast<long>((currentTime.QuadPart - previousTime.QuadPart) * 1000.0 / frequency.QuadPart);
+
+		float updateLength = static_cast<float>((currentTime.QuadPart - previousTime.QuadPart) * 1000) / static_cast<float>(frequency.QuadPart);
 		
 		//Previous time = current time
 		previousTime = currentTime;
@@ -58,13 +59,10 @@ void Game::gameLoop(GameStateManager* gsm) {
 		//Increase FPS since one frame has passed
 		fps++;
 
-		//gsm->updateGameTime(updateLength);
 
 		//If lastFpsTime > 1000 ms, set FPS to 0 and start re-calculating
 		if (lastFpsTime >= 1000) {
 			gsm->setFps(fps);
-			
-			//std::cout << "FPS: "<< fps << std::endl;
 			lastFpsTime = 0;
 			fps = 0;
 		}
